requirements.txt
tabulate
pymaybe
biopython
boto3
botocore